sales/requirements.txt
pytesseract
Pillow
opencv-python
pandas
python-dotenv
openpyxl
tabulate
watchdog
